Xnest: drop hack for building 32bit server against 64bit libs
Xnest.h contains a weird historical hack for building 32bit server with 64bit
libraries. It redeclares a bunch of stanard X types (eg. from Xdefs.h) to
32 bit size. Those hacks are ugly to maintain and can easily cause trouble
(eg. if include order gets mixed up). And it's not at all needed anyways.

Those kind of situations are easily solved, in a much more robust way:
just run build the build it in a chroot or container, or do a pretty
boring usual crosscompile.

No need to carry special hacks for things that already done out-of-the-box
by proper tooling / build environment.

Signed-off-by: Enrico Weigelt, metux IT consult <info@metux.net>