os: drop conditional build on TCPCONN define
This symbol is always defined for aeons, at least since meson transition,
and there doesn't seem any desire to ever switch it off again, so we
can drop the #ifdef's.

Note that the symbol still needs to be defined for xtrans, but doing so
in os/meson.build should be enough.

Signed-off-by: Enrico Weigelt, metux IT consult <info@metux.net>