meson_options: set default for HAL integration to false
This should be a deliberate decision of the user/distro. Just the presense
of some of its libraries in the build environment doesn't automatically
mean the user/distro actually wants it to be used.

HAL is pretty much obsolete, so we're considering to drop it entirely.
Anybody who's still needs it should call out loud now.

Signed-off-by: Enrico Weigelt, metux IT consult <info@metux.net>